instagram-scraper/requirements.txt
# Instagram Trending Hashtag Discovery - Dependencies
# Python 3.10+ required
#
# Installation:
#   1. pip install -r requirements.txt
#   2. playwright install chromium
#   3. python -c "from textblob import TextBlob; TextBlob('test').sentiment"  # Download TextBlob corpora

# ============================================================================
# Core Dependencies (Required)
# ============================================================================

# Browser Automation - Used for Instagram scraping
playwright>=1.40.0,<2.0.0

# Database Client - Supabase integration
supabase>=2.0.2,<3.0.0

# Task Scheduling - Automated runs
apscheduler>=3.10.4,<4.0.0

# ============================================================================
# Analysis Dependencies (Required)
# ============================================================================

# Sentiment Analysis - TextBlob for sentiment detection
textblob>=0.17.1,<1.0.0

# Advanced Sentiment Analysis - VADER for enhanced sentiment (optional but recommended)
vaderSentiment>=3.3.2,<4.0.0

# Language Detection - Automatic language identification
langdetect>=1.0.9,<2.0.0

# ============================================================================
# Utility Dependencies (Required)
# ============================================================================



# Environment Variables - Load .env files
python-dotenv>=1.0.0,<2.0.0


# ============================================================================
# Data Processing (Optional - for future enhancements)
# ============================================================================

# Data Analysis - DataFrame operations (if needed for data processing)
pandas>=2.1.3,<3.0.0

# Numerical Computing - Mathematical operations (dependency of pandas)
numpy>=1.24.3,<2.0.0

# ============================================================================
# Testing Dependencies (Required for running tests)
# ============================================================================

# Testing Framework - Unit testing
pytest>=7.4.0,<8.0.0

# Mocking Support - Enhanced mocking capabilities
pytest-mock>=3.11.1,<4.0.0

# Coverage Reporting - Code coverage analysis
pytest-cov>=4.1.0,<5.0.0

# ============================================================================
# Observability Dependencies (Required for structured logging and metrics)
# ============================================================================

# Context Variables - For request/trace ID tracking (built-in Python 3.7+)
# No external dependency needed - contextvars is in stdlib

# ============================================================================
# Configuration API Dependencies (Required for dynamic configuration)
# ============================================================================

# Web Framework - REST API for configuration management
flask>=3.0.0,<4.0.0

# CORS Support - Enable cross-origin requests for admin UI
flask-cors>=4.0.0,<5.0.0